Two fixes for memory leaks, the second is critical. 1. Increased the array sizes for several character strings to accommodate when they get concatenated. This fixes several compilation-time warnings. Likely these bufferes would not be overrun, but at least now they are protected against it. 2. Critical fix to ThreadPool.h that was causing a memory leak during clean-up after Comet analysis completion. This was causing an intermittent seg fault 11 on Linux runs. Perhaps similar problems elsewhere. It looks like CometSearchManager.cpp had some indirect code that ineffectively dealt with the issue (see approx. line 595) but I'm not sure. Further testing is needed.

+++ CometSearch/CometSearch.cpp
@@ -249,7 +249,7 @@ bool CometSearch::RunSearch(int iPercentStart,
       int  iLenAttributeVariantComplex = (int)strlen(szPeffAttributeVariantComplex);
       int  iLenAttributeMod = (int)strlen(szPeffAttributeMod);
       int  iNumBadChars = 0; // count # of bad (non-printing) characters in header 
-
+ 
       bool bHeadOfFasta = true;
       // Loop through entire database.
       while(!feof(fp))
@@ -1579,7 +1579,7 @@ bool CometSearch::SearchForPeptides(struct sDBEntry dbe,
    if (_proteinInfo.iPeffNewResidueCount>0) // if not a deletion, adjust the position by 1
       iPeffRequiredVariantPositionB--;
 
-   if (iPeffRequiredVariantPosition == strlen(szProteinSeq)) //Edge case where PEFF variant is simply a truncation of the c-terminal amino acids
+   if (iPeffRequiredVariantPosition == (int)strlen(szProteinSeq)) //Edge case where PEFF variant is simply a truncation of the c-terminal amino acids
    {
       iPeffRequiredVariantPosition--;
       iPeffRequiredVariantPositionB--;
